Refactored message matching.

+++ PIOsimHD-Simulator/src/de/hd/pvs/piosim/simulator/program/Allreduce/BinaryTree.java
@@ -43,7 +43,7 @@ public class BinaryTree
 {
 	
 	@Override
-	public void process( Allreduce cmd, CommandProcessing OUTresults, GClientProcess client, int step, NetworkJobs compNetJobs) {
+	public void process( Allreduce cmd, CommandProcessing OUTresults, GClientProcess client, long step, NetworkJobs compNetJobs) {
 
 		if (cmd.getCommunicator().getSize() == 1){
 			// finished ...
