+++ PIOsimHD-Simulator/src/de/hd/pvs/piosim/simulator/program/FileClose/SimpleClose.java
@@ -38,7 +38,7 @@ extends CommandImplementation<de.hd.pvs.piosim.model.program.commands.Fileclose>
 {
 	@Override
 	public void process(Fileclose cmd, CommandProcessing OUTresults,
-			GClientProcess client, int step, NetworkJobs compNetJobs) {
+			GClientProcess client, long step, NetworkJobs compNetJobs) {
 		if (step == CommandProcessing.STEP_START){
 			Barrier barrier = new Barrier();
 			barrier.setCommunicator(cmd.getCommunicator());
